+++ playlist_bridge/__init__.py
@@ -1,5 +1,5 @@
 """Playlist Bridge package."""
 
-__version__ = "2.1.0"
+__version__ = "2.2.0-beta.1"
 
-__build__ = "20260918.9"
+__build__ = "20260919.1"